+++ Example/cauli-ios-example/Pods/Target Support Files/Cauli/Cauli-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Cauli : NSObject
-@end
-@implementation PodsDummy_Cauli
-@end