+++ Example/cauli-ios-example/Pods/Target Support Files/Cauliframework/Cauliframework-umbrella.h
@@ -10,7 +10,8 @@
 #endif
 #endif
 
+#import "Cauli.h"
 
-FOUNDATION_EXPORT double Pods_CauliTestsVersionNumber;
-FOUNDATION_EXPORT const unsigned char Pods_CauliTestsVersionString[];
+FOUNDATION_EXPORT double CauliframeworkVersionNumber;
+FOUNDATION_EXPORT const unsigned char CauliframeworkVersionString[];