+++ Pods/Target Support Files/Pods-Cauli/Pods-Cauli-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_Cauli : NSObject
-@end
-@implementation PodsDummy_Pods_Cauli
-@end