+++ Pods/Target Support Files/Pods-Cauliframework/Pods-Cauliframework-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Pods_Cauliframework : NSObject
+@end
+@implementation PodsDummy_Pods_Cauliframework
+@end